Separated Routes and handler from Main
Everything is in the router folder in different file
